+++ sdk/communication/communication-sms/src/optOutsClient.ts
@@ -8,31 +8,83 @@ import { generateOptOutRequest } from "./utils/smsUtils.js";
 import { extractOperationOptions } from "./extractOperationOptions.js";
 import type { OperationOptions } from "@azure/core-client";
 
+/**
+ * The result of Opt Out Check request.
+ */
 export interface OptOutCheckResult {
+   /**
+   * The recipient's phone number in E.164 format.
+   */
   to: string;
+   /**
+   * Indicates if the recipient's phone number in opted out from receiving messages or not.
+   */
   isOptedOut: boolean;
+   /**
+   * HTTP Status code.
+   */
   httpStatusCode: number;
+  /**
+   * Optional error message in case of 4xx/5xx/repeatable errors.
+   */
   errorMessage?: string;
 }
 
+/**
+ * The result of Opt Out Add request.
+ */
 export interface OptOutAddResult {
+  /**
+   * The recipient's phone number in E.164 format.
+   */
   to: string;
+   /**
+   * HTTP Status code.
+   */
   httpStatusCode: number;
+  /**
+   * Optional error message in case of 4xx/5xx/repeatable errors.
+   */
   errorMessage?: string;
 }
 
+/**
+ * The result of Opt Out Remove request.
+ */
 export interface OptOutRemoveResult {
+  /**
+   * The recipient's phone number in E.164 format.
+   */
   to: string;
+    /**
+   * HTTP Status code.
+   */
   httpStatusCode: number;
+  /**
+   * Optional error message in case of 4xx/5xx/repeatable errors.
+   */
   errorMessage?: string;
 }
 
+/**
+ * Client options used to configure OptOuts Client API Check requests.
+ */
 export interface CheckOptions extends OperationOptions {}
 
+/**
+ * Client options used to configure OptOuts Client API Add requests.
+ */
 export interface AddOptions extends OperationOptions {}
 
+/**
+ * Client options used to configure OptOuts Client API Remove requests.
+ */
 export interface RemoveOptions extends OperationOptions {}
 
+/**
+ * A OptOutsClient represents a Client to the Azure Communication Sms service allowing you
+ * to call Opt Out Management Api methods.
+ */
 export class OptOutsClient {
   private readonly api: SmsApiClient;
 
@@ -40,6 +92,13 @@ export class OptOutsClient {
     this.api = api;
   }
 
+   /**
+   * Removes phone numbers from the optouts list. 
+   *
+   * @param from - The sender's phone number
+   * @param to - The recipient's phone numbers
+   * @param options - Additional request options
+   */
   public async remove(
     from: string,
     to: string[],
@@ -59,6 +118,13 @@ export class OptOutsClient {
     });
   }
 
+   /**
+   * Adds phone numbers from the optouts list. 
+   *
+   * @param from - The sender's phone number
+   * @param to - The recipient's phone numbers
+   * @param options - Additional request options
+   */
   public async add(
     from: string,
     to: string[],
@@ -78,6 +144,13 @@ export class OptOutsClient {
     });
   }
 
+   /**
+   * Checks if phone numbers are in the optouts list. 
+   *
+   * @param from - The sender's phone number
+   * @param to - The recipient's phone numbers
+   * @param options - Additional request options
+   */
   public async check(
     from: string,
     to: string[],