+++ sdk/agrifood/agrifood-farming-rest/src/index.ts
@@ -1,17 +1,17 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import FarmBeats from "./farmBeats";
+import FarmBeats from "./farmBeats.js";
 
-export * from "./farmBeats";
-export * from "./parameters";
-export * from "./responses";
-export * from "./clientDefinitions";
-export * from "./isUnexpected";
-export * from "./models";
-export * from "./outputModels";
-export * from "./paginateHelper";
-export * from "./pollingHelper";
-export * from "./serializeHelper";
+export * from "./farmBeats.js";
+export * from "./parameters.js";
+export * from "./responses.js";
+export * from "./clientDefinitions.js";
+export * from "./isUnexpected.js";
+export * from "./models.js";
+export * from "./outputModels.js";
+export * from "./paginateHelper.js";
+export * from "./pollingHelper.js";
+export * from "./serializeHelper.js";
 
 export default FarmBeats;