+++ sdk/cognitivelanguage/ai-language-conversations/src/conversationAnalysisClient.ts
@@ -14,16 +14,16 @@ import type {
   ConversationAnalysisClientOptionalParams,
   ConversationAnalysisOptionalParams,
   ConversationAnalysisResponse,
-} from "./models";
-import { DEFAULT_COGNITIVE_SCOPE, SDK_VERSION } from "./constants";
+} from "./models.js";
+import { DEFAULT_COGNITIVE_SCOPE, SDK_VERSION } from "./constants.js";
 import type { KeyCredential, TokenCredential } from "@azure/core-auth";
 import { isTokenCredential } from "@azure/core-auth";
 import type { PollOperationState, PollerLike } from "@azure/core-lro";
 import type { TracingClient } from "@azure/core-tracing";
 import { createTracingClient } from "@azure/core-tracing";
-import { ConversationAnalysisClient as GeneratedClient } from "./generated";
+import { ConversationAnalysisClient as GeneratedClient } from "./generated/index.js";
 import { bearerTokenAuthenticationPolicy } from "@azure/core-rest-pipeline";
-import { conversationAnalysisAzureKeyCredentialPolicy } from "./azureKeyCredentialPolicy";
+import { conversationAnalysisAzureKeyCredentialPolicy } from "./azureKeyCredentialPolicy.js";
 
 /**
  * A client for interacting with the conversational language understanding