+++ sdk/easm/defender-easm-rest/test/public/discoTemplatesTest.spec.ts
@@ -3,12 +3,11 @@
 
 import type { Recorder } from "@azure-tools/test-recorder";
 import { assertEnvironmentVariable } from "@azure-tools/test-recorder";
-import { assert } from "chai";
-import { createRecorder } from "./utils/recordedClient";
-import type { Context } from "mocha";
-import type { EasmClient } from "../../src";
-import EasmDefender, { isUnexpected } from "../../src";
+import { createRecorder } from "./utils/recordedClient.js";
+import type { EasmClient } from "../../src/index.js";
+import EasmDefender, { isUnexpected } from "../../src/index.js";
 import { createTestCredential } from "@azure-tools/test-credential";
+import { describe, it, assert, beforeEach, afterEach } from "vitest";
 
 describe("Discovery Templates Test", () => {
   let recorder: Recorder;
@@ -16,8 +15,8 @@ describe("Discovery Templates Test", () => {
   let template_id: string;
   let partial_name: string;
 
-  beforeEach(async function (this: Context) {
-    recorder = await createRecorder(this);
+  beforeEach(async (ctx) => {
+    recorder = await createRecorder(ctx);
     const subscription_id = assertEnvironmentVariable("SUBSCRIPTION_ID");
     const resource_group = assertEnvironmentVariable("RESOURCEGROUPNAME");
     const workspace_name = assertEnvironmentVariable("WORKSPACENAME");
@@ -35,7 +34,7 @@ describe("Discovery Templates Test", () => {
     partial_name = "ku";
   });
 
-  afterEach(async function () {
+  afterEach(async () => {
     await recorder.stop();
   });