+++ sdk/identity/identity/eslint.config.mjs
@@ -1,18 +1,18 @@
 import azsdkEslint from "@azure/eslint-plugin-azure-sdk";
 
-export default [
+export default azsdkEslint.config([
   { ignores: ["test/manual*"] },
-  ...azsdkEslint.configs.recommended,
   {
     rules: {
       "@azure/azure-sdk/ts-naming-options": "warn",
     },
   },
   {
-    files: ["test/snippets.spec.ts"],
-    rules: {
-      "@typescript-eslint/ban-ts-comment": "off",
-      "@typescript-eslint/no-unused-vars": "off",
+    files: ["**/*.ts", "**/*.cts", "**/*.mts"],
+    languageOptions: {
+      parserOptions: {
+        project: ["./tsconfig.test.json"],
+      },
     },
   },
-];
+]);