+++ sdk/monitor/monitor-query/test/internal/unit/modelConverters.unittest.spec.ts
@@ -244,7 +244,6 @@ describe("Model unit tests", () => {
         // NOTE: _response is not returned as part of our track 2 response.
       };
 
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const { getMetricByName, ...rest } = actualConvertedResponse;
       assert.deepEqual(
         { ...rest } as Omit<MetricsQueryResult, "getMetricByName">,