+++ sdk/synapse/synapse-access-control/eslint.config.mjs
@@ -1,7 +1,6 @@
 import azsdkEslint from "@azure/eslint-plugin-azure-sdk";
 
-export default [
-  ...azsdkEslint.configs.recommended,
+export default azsdkEslint.config([
   {
     files: ["src/**/*.ts"],
     rules: {
@@ -9,4 +8,4 @@ export default [
       "@azure/azure-sdk/ts-use-interface-parameters": "off",
     },
   },
-];
+]);