+++ sdk/test-utils/test-utils-vitest/eslint.config.mjs
@@ -2,4 +2,12 @@ import azsdkEslint from "@azure/eslint-plugin-azure-sdk";
 
 export default [
   ...azsdkEslint.configs.internal,
+  {
+    files: ["**/*.ts", "**/*.cts", "**/*.mts"],
+    languageOptions: {
+      parserOptions: {
+        project: ["./tsconfig.test.json"],
+      },
+    },
+  },
 ];