+++ sdk/communication/communication-sms/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Added support for Opt Out Management Api to:
   - Opt-out the recipient.
   - Opt-in the recipient.
-  - Check if the recpient is opted-out or not.
+  - Check if the recipient is opted-out or not.
 
 ## 1.2.0-beta.1 (2024-05-02)